11/06/2022-01:11

+++ src/main/java/fr/fms/dao/CategoryRepository.java
@@ -4,15 +4,15 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import fr.fms.entities.Article;
 import fr.fms.entities.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+// public interface CategoryRepository extends JpaRepository<Category, Long> 
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
 	public List<Category> findById(int id);
 
 	public List<Category> findByName(String name);
-	
+
 	public List<Category> findAllByOrderByIdAsc();
 
 	public List<Category> findAllByOrderByIdDesc();
